Refactor ShapesetGenerator & ShapesetGeneratorTest: they are now in .shapesetgenerator packages
